+++ src/untangled/client/impl/data_fetch.cljs
@@ -331,9 +331,9 @@
                         (swap! app-state assoc :untangled/server-error error)
                         (doseq [item loading-items]
                           (swap! app-state (fn [s]
-                                             (-> s
-                                                 (update :untangled/loads-in-progress disj (data-uuid item))
-                                                 (update-in (data-path item) set-failed! error))))))
+                                             (cond-> s
+                                               (data-marker? item) (update-in (data-path item) set-failed! error)
+                                               (update :untangled/loads-in-progress disj (data-uuid item)))))))
           run-fallbacks (fn [] (doseq [item loading-items]
                                  (when-let [fallback-symbol (::fallback item)]
                                    (reset! ran-fallbacks true)