+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 -----
 - Fixed bug with global-error-callback not being called with a server error returns no body
 - Fixed bug that prevents error processing if the server is completely down
+- Added reset-history function to reset UntangledApplication Om cache history.
 
 0.5.5
 -----